chore: organize minishell.h, update Doxygen comments
- Alphabetized minishell.h by directory
- Updated Doxygen comments for existing functions
- Added concise comments to new AST build functions
- Ensured all functions comply with Norminette (<25 lines each)
