+++ src/parser/ast_build.c
@@ -1,27 +1,20 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ast_build.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: diade-so <diade-so@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/10/07 10:21:04 by diade-so          #+#    #+#             */
-/*   Updated: 2025/10/07 16:36:44 by diade-so         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
 #include "minishell.h"
-//t_token *find_first_pipe(t_token *tokens) TODO:DELETE
+
+/**
+ * @brief Find the first pipe token in a token range.
+ *
+ * Scans the tokens from `start` up to (but not including) `end`
+ * and returns the first token representing a pipe (`|`).
+ *
+ * @param start Pointer to the first token in the range.
+ * @param end Pointer to the token marking the end of the range (excluded).
+ * @return Pointer to the first pipe token, or NULL if none found.
+ */
 t_token	*find_first_pipe(t_token *start, t_token *end)
 {
 	t_token	*curr;
 
-	//curr = tokens;TODO:DELETE
 	curr = start;
-	// while (curr) TODO: DELETE
-	// conditions now include  end pointer, which is necessary since
-	// we're no longer using next, there's no way to detect NULL;
-	// instead the end pointer tells us when to stop.
 	while (curr && curr != end)
 	{
 		if (curr->op_type == OP_PIPE)
@@ -31,57 +24,56 @@ t_token	*find_first_pipe(t_token *start, t_token *end)
 	return (NULL);
 }
 
-/* TODO: DELETE ENTIRELY
-t_token	*split_at_pipe(t_token *tokens, t_token *pipe_token)
+/**
+ * @brief Free command resources safely.
+ *
+ * Frees either a command node or its argv array, depending on what
+ * was allocated.
+ *
+ * @param cmd_node Command AST node to free (can be NULL).
+ * @param argv Argument array to free if no node exists.
+ * @return Always returns NULL for convenience in cleanup chains.
+ */
+static t_ast	*cleanup_command(t_ast *cmd_node, char **argv)
 {
-	t_token	*curr;
-	t_token *right_subtree;
+	if (cmd_node)
+		free_ast(cmd_node);
+	else if (argv)
+		free_strings_array(argv);
+	return (NULL);
+}
 
-	curr = tokens;
-	// split the token BEFORE the pipe and break the chain
-	while (curr && curr->next != pipe_token)
-		curr = curr->next;
-	if (curr)
-		curr->next = NULL; //break the chain before the pipe
-	//return tokens AFTER the pipe (the right side)
-	return (right_subtree);
-}*/
-
-//t_ast	*build_simple_command(t_token *tokens) TODO:DELETE
+/**
+ * @brief Build a simple command node from tokens.
+ *
+ * Creates a command node with its argv and attached redirections.
+ * Returns NULL on allocation failure or invalid redirection setup.
+ *
+ * @param start Start of token range for the command.
+ * @param end End of token range (excluded).
+ * @return A fully built command node, or NULL on error.
+ */
 t_ast	*build_simple_command(t_token *start, t_token *end)
 {
 	char	**argv;
 	t_ast	*cmd_node;
 	t_ast	*redir_head;
 
-//	if (!tokens) TODO: DELETE
 	if (!start || start == end)
 		return (NULL);
-//	argv = collect_argv(tokens); TODO:DELETE
 	argv = collect_argv(start, end);
 	if (!argv)
 		return (NULL);
 	cmd_node = NULL;
-	// create cmd_node if we have arguments
-	if (argv && argv[0]) // check if there is at least one arg
+	if (argv && argv[0])
 	{
 		cmd_node = create_cmd_node(argv);
-		{
-			if (!cmd_node)
-				return (free_strings_array(argv), NULL);
-		}
+		if (!cmd_node)
+			return (free_strings_array(argv), NULL);
 	}
-	// collect all redirections into a chain
-//	redir_head = collect_redirections(tokens); TODO:delete
 	redir_head = collect_redirections(start, end);
 	if (!redir_head && has_redirections(start, end))
-	{
-		if (cmd_node)
-			free_ast(cmd_node);
-		else
-			free_strings_array(argv);
-		return (NULL);
-	}
+		return (cleanup_command(cmd_node, argv));
 	if (cmd_node)
 	{
 		cmd_node->right = redir_head;
@@ -90,6 +82,16 @@ t_ast	*build_simple_command(t_token *start, t_token *end)
 	return (redir_head);
 }
 
+/**
+ * @brief Recursively build an AST from a token range.
+ *
+ * Splits tokens by the first pipe symbol and builds left/right subtrees.
+ * Base case: a simple command with optional redirections.
+ *
+ * @param start Start of token range.
+ * @param end End of token range (NULL for full list).
+ * @return Root of the constructed AST, or NULL on failure.
+ */
 static t_ast	*build_ast_recursive(t_token *start, t_token *end)
 {
 	t_token	*pipe_token;
@@ -97,40 +99,17 @@ static t_ast	*build_ast_recursive(t_token *start, t_token *end)
 	t_ast	*right_subtree;
 	t_ast	*pipe_node;
 
-	// ADDED: base case empty list
 	if (!start || start == end)
 		return (NULL);
-	
-	// pipe_token = find_first_pipe(tokens); TODO: DELETE!
-	// find left most pipe in the segment
-	// the segment now has clear start and end points.
 	pipe_token = find_first_pipe(start, end);
-
-	// if no pipe found build simple cmd
 	if (!pipe_token)
 		return (build_simple_command(start, end));
-
-	// right_tokens = split_at_pipe(tokens, pipe_token); TODO: DELETE!
-	// this is where it's breaking the tokens list making it impossible
-	// for us to free the tokens list entirely resulting in leaks
-	// for all nodes right of the pipe. we no longer need this
-
-	//recursively build subtrees
 	left_subtree = build_simple_command(start, pipe_token);
 	if (!left_subtree)
 		return (NULL);
-
-
-	// DELETE: right_subtree = build_ast_from_tokens(right_tokens);
-	
-	
 	right_subtree = build_ast_recursive(pipe_token->next, end);
 	if (!right_subtree)
 		return (free_ast(left_subtree), NULL);
-	// now that the tree is no longer broken, we can start the right
-	// subtree at pipe_token->next until end pointer.
-
-
 	pipe_node = create_pipe_node(left_subtree, right_subtree);
 	if (!pipe_node)
 	{
@@ -141,11 +120,15 @@ static t_ast	*build_ast_recursive(t_token *start, t_token *end)
 	return (pipe_node);
 }
 
-// new wrapper function, since there is now a start and end,
-// the wrapper function is necessary to update those pointers.
-t_ast *build_ast_from_tokens(t_token *tokens)
+/**
+ * @brief Entry point for AST construction.
+ *
+ * Builds a full AST from the given token list.
+ *
+ * @param tokens Head of the token list.
+ * @return Root of the AST, or NULL on failure.
+ */
+t_ast	*build_ast_from_tokens(t_token *tokens)
 {
-	// here we set the start and end tokens.
-	// start: tokens, end: NULL
-	return  (build_ast_recursive(tokens, NULL));
+	return (build_ast_recursive(tokens, NULL));
 }
